refactor(global): QueueStatus, QueuePhase 내부 클래스를 독립 파일로 분리
- `EventQueueManager`의 내부 record/enum을 `global.queue` 패키지 최상위 타입으로 추출
- `QueueStatusResponse`, 테스트 클래스의 import 경로 수정
