+++ src/main/java/com/example/openticket/global/queue/QueuePhase.java
@@ -1,0 +1,5 @@
+package com.example.openticket.global.queue;
+
+public enum QueuePhase {
+    WAITING, ALLOWED
+}
